requirements.txt
ipython
jupyterlab
matplotlib
notebook
numpy
pandas
pip
python-dotenv
scikit-learn
snowflake-connector-python
requests
os
subprocess
pathlib
json